# Conflicts:
#	src/api/client/rpc_connection.rs
#	src/api/client/session.rs
#	src/signalling/room.rs

+++ src/signalling/room_service.rs
@@ -7,8 +7,10 @@ use actix::{
 };
 use derive_more::Display;
 use failure::Fail;
-use futures::future::{self, Future};
-use medea_control_api_proto::grpc::api::Element as ElementProto;
+use futures::future::{
+    self, FutureExt as _, LocalBoxFuture, TryFutureExt as _,
+};
+use medea_control_api_proto::grpc::api as proto;
 
 use crate::{
     api::control::{
@@ -141,7 +143,7 @@ impl RoomService {
     fn close_room(
         &self,
         id: RoomId,
-    ) -> Box<dyn Future<Item = (), Error = MailboxError>> {
+    ) -> LocalBoxFuture<'static, Result<(), MailboxError>> {
         if let Some(room) = self.room_repo.get(&id) {
             shutdown::unsubscribe(
                 &self.graceful_shutdown,
@@ -150,13 +152,17 @@ impl RoomService {
             );
 
             let room_repo = self.room_repo.clone();
-
-            Box::new(room.send(Close).map(move |_| {
-                debug!("Room [id = {}] removed.", id);
-                room_repo.remove(&id);
-            }))
+            let sending = room.send(Close);
+            async move {
+                let res = sending.await;
+                if res.is_ok() {
+                    room_repo.remove(&id);
+                }
+                res
+            }
+            .boxed_local()
         } else {
-            Box::new(futures::future::ok(()))
+            async { Ok(()) }.boxed_local()
         }
     }
 
@@ -287,7 +293,7 @@ pub struct CreateMemberInRoom {
 }
 
 impl Handler<CreateMemberInRoom> for RoomService {
-    type Result = ResponseFuture<Sids, RoomServiceError>;
+    type Result = ResponseFuture<Result<Sids, RoomServiceError>>;
 
     fn handle(
         &mut self,
@@ -300,17 +306,17 @@ impl Handler<CreateMemberInRoom> for RoomService {
         sids.insert(msg.id.to_string(), sid);
 
         if let Some(room) = self.room_repo.get(&room_id) {
-            Box::new(
-                room.send(CreateMember(msg.id, msg.spec))
-                    .map_err(RoomServiceError::RoomMailboxErr)
-                    .and_then(move |r| {
-                        r.map_err(RoomServiceError::from).map(move |_| sids)
-                    }),
-            )
+            let sending = room.send(CreateMember(msg.id, msg.spec));
+            async {
+                sending.await.map_err(RoomServiceError::RoomMailboxErr)??;
+                Ok(sids)
+            }
+            .boxed_local()
         } else {
-            Box::new(future::err(RoomServiceError::RoomNotFound(
-                Fid::<ToRoom>::new(room_id),
+            future::err(RoomServiceError::RoomNotFound(Fid::<ToRoom>::new(
+                room_id,
             )))
+            .boxed_local()
         }
     }
 }
@@ -327,7 +333,7 @@ pub struct CreateEndpointInRoom {
 }
 
 impl Handler<CreateEndpointInRoom> for RoomService {
-    type Result = ResponseFuture<Sids, RoomServiceError>;
+    type Result = ResponseFuture<Result<Sids, RoomServiceError>>;
 
     fn handle(
         &mut self,
@@ -338,21 +344,21 @@ impl Handler<CreateEndpointInRoom> for RoomService {
         let endpoint_id = msg.id;
 
         if let Some(room) = self.room_repo.get(&room_id) {
-            Box::new(
-                room.send(CreateEndpoint {
-                    member_id,
-                    endpoint_id,
-                    spec: msg.spec,
-                })
-                .map_err(RoomServiceError::RoomMailboxErr)
-                .and_then(|r| {
-                    r.map_err(RoomServiceError::from).map(|_| HashMap::new())
-                }),
-            )
+            let sending = room.send(CreateEndpoint {
+                member_id,
+                endpoint_id,
+                spec: msg.spec,
+            });
+            async {
+                sending.await.map_err(RoomServiceError::RoomMailboxErr)??;
+                Ok(HashMap::new())
+            }
+            .boxed_local()
         } else {
-            Box::new(future::err(RoomServiceError::RoomNotFound(
-                Fid::<ToRoom>::new(room_id),
+            future::err(RoomServiceError::RoomNotFound(Fid::<ToRoom>::new(
+                room_id,
             )))
+            .boxed_local()
         }
     }
 }
@@ -385,6 +391,11 @@ impl DeleteElements<Unvalidated> {
 
     /// Validates request. It must have at least one fid, all fids must share
     /// same [`RoomId`].
+    ///
+    /// # Errors
+    ///
+    /// Errors with [`RoomServiceError::EmptyUrisList`] if [`DeleteElements`]
+    /// consists of an empty [`Vec`] of [`StatefulFid`]s.
     pub fn validate(
         self,
     ) -> Result<DeleteElements<Validated>, RoomServiceError> {
@@ -432,7 +443,7 @@ pub struct DeleteElements<T> {
 }
 
 impl Handler<DeleteElements<Validated>> for RoomService {
-    type Result = ResponseFuture<(), RoomServiceError>;
+    type Result = ResponseFuture<Result<(), RoomServiceError>>;
 
     // TODO: delete 'clippy::unnecessary_filter_map` when drain_filter TODO will
     //       be resolved.
@@ -443,48 +454,48 @@ impl Handler<DeleteElements<Validated>> for RoomService {
         _: &mut Self::Context,
     ) -> Self::Result {
         let mut deletes_from_room: Vec<StatefulFid> = Vec::new();
+
         // TODO: use Vec::drain_filter when it will be in stable
-        let room_messages_futs: Vec<
-            Box<dyn Future<Item = (), Error = MailboxError>>,
-        > = msg
-            .fids
-            .into_iter()
-            .filter_map(|fid| {
-                if let StatefulFid::Room(room_id) = fid {
-                    Some(self.close_room(room_id.take_room_id()))
-                } else {
-                    deletes_from_room.push(fid);
-                    None
-                }
-            })
-            .collect();
+        let room_messages_futs: Vec<ResponseFuture<Result<(), MailboxError>>> =
+            msg.fids
+                .into_iter()
+                .filter_map(|fid| {
+                    if let StatefulFid::Room(room_id) = fid {
+                        Some(self.close_room(room_id.take_room_id()))
+                    } else {
+                        deletes_from_room.push(fid);
+                        None
+                    }
+                })
+                .collect();
 
         if !room_messages_futs.is_empty() {
-            Box::new(
-                futures::future::join_all(room_messages_futs)
-                    .map(|_| ())
-                    .map_err(RoomServiceError::RoomMailboxErr),
-            )
+            future::try_join_all(room_messages_futs)
+                .map_ok(|_| ())
+                .map_err(RoomServiceError::RoomMailboxErr)
+                .boxed_local()
         } else if !deletes_from_room.is_empty() {
             let room_id = deletes_from_room[0].room_id().clone();
 
             if let Some(room) = self.room_repo.get(&room_id) {
-                Box::new(
-                    room.send(Delete(deletes_from_room))
-                        .map_err(RoomServiceError::RoomMailboxErr),
-                )
+                let sending = room.send(Delete(deletes_from_room));
+                async {
+                    sending.await.map_err(RoomServiceError::RoomMailboxErr)?;
+                    Ok(())
+                }
+                .boxed_local()
             } else {
-                Box::new(future::ok(()))
+                future::ok(()).boxed_local()
             }
         } else {
-            Box::new(future::err(RoomServiceError::EmptyUrisList))
+            future::err(RoomServiceError::EmptyUrisList).boxed_local()
         }
     }
 }
 
 /// Serialized to protobuf `Element`s which will be returned from [`Get`] on
 /// success result.
-type SerializedElements = HashMap<StatefulFid, ElementProto>;
+type SerializedElements = HashMap<StatefulFid, proto::Element>;
 
 /// Message which returns serialized to protobuf objects by provided
 /// [`Fid`].
@@ -493,7 +504,7 @@ type SerializedElements = HashMap<StatefulFid, ElementProto>;
 pub struct Get(pub Vec<StatefulFid>);
 
 impl Handler<Get> for RoomService {
-    type Result = ResponseFuture<SerializedElements, RoomServiceError>;
+    type Result = ResponseFuture<Result<SerializedElements, RoomServiceError>>;
 
     fn handle(&mut self, msg: Get, _: &mut Self::Context) -> Self::Result {
         let mut rooms_elements = HashMap::new();
@@ -506,9 +517,8 @@ impl Handler<Get> for RoomService {
                     .or_insert_with(Vec::new)
                     .push(fid);
             } else {
-                return Box::new(future::err(RoomServiceError::RoomNotFound(
-                    fid.into(),
-                )));
+                return future::err(RoomServiceError::RoomNotFound(fid.into()))
+                    .boxed_local();
             }
         }
 
@@ -517,20 +527,21 @@ impl Handler<Get> for RoomService {
             futs.push(room.send(SerializeProto(elements)));
         }
 
-        Box::new(
-            futures::future::join_all(futs)
-                .map_err(RoomServiceError::RoomMailboxErr)
-                .and_then(|results| {
-                    let mut all = HashMap::new();
-                    for result in results {
-                        match result {
-                            Ok(res) => all.extend(res),
-                            Err(e) => return Err(RoomServiceError::from(e)),
-                        }
-                    }
-                    Ok(all)
-                }),
-        )
+        async {
+            let results = future::try_join_all(futs)
+                .await
+                .map_err(RoomServiceError::RoomMailboxErr)?;
+
+            let mut all = HashMap::new();
+            for res in results {
+                match res {
+                    Ok(r) => all.extend(r),
+                    Err(e) => return Err(RoomServiceError::from(e)),
+                }
+            }
+            Ok(all)
+        }
+        .boxed_local()
     }
 }
 
@@ -563,7 +574,7 @@ mod delete_elements_validation_specs {
         let mut elements = DeleteElements::new();
         ["room_id/member", "another_room_id/member"]
             .iter()
-            .map(|fid| StatefulFid::try_from(fid.to_string()).unwrap())
+            .map(|fid| StatefulFid::try_from((*fid).to_string()).unwrap())
             .for_each(|fid| elements.add_fid(fid));
 
         match elements.validate() {
@@ -593,7 +604,7 @@ mod delete_elements_validation_specs {
             "room_id/member_id/endpoint_id",
         ]
         .iter()
-        .map(|fid| StatefulFid::try_from(fid.to_string()).unwrap())
+        .map(|fid| StatefulFid::try_from((*fid).to_string()).unwrap())
         .for_each(|fid| elements.add_fid(fid));
 
         assert!(elements.validate().is_ok());
@@ -666,50 +677,39 @@ mod room_service_specs {
             $create_msg:expr,
             $element_fid:expr,
             $test:expr
-        ) => {{
-            let get_msg = Get(vec![$element_fid.clone()]);
-            $room_service
-                .send($create_msg)
-                .and_then(move |res| {
-                    res.unwrap();
-                    $room_service.send(get_msg)
-                })
-                .map(move |r| {
-                    let mut resp = r.unwrap();
-                    resp.remove(&$element_fid).unwrap()
-                })
-                .map($test)
-                .map(|_| actix::System::current().stop())
-                .map_err(|e| panic!("{:?}", e))
-        }};
+        ) => {
+            async move {
+                let get_msg = Get(vec![$element_fid.clone()]);
+                $room_service.send($create_msg).await.unwrap().unwrap();
+                let mut resp =
+                    $room_service.send(get_msg).await.unwrap().unwrap();
+                resp.remove(&$element_fid).unwrap();
+                actix::System::current().stop();
+            }
+        };
     }
 
-    #[test]
-    fn create_room() {
-        let sys = actix::System::new("room-service-tests");
-
+    #[actix_rt::test]
+    async fn create_room() {
         let room_service = room_service(RoomRepository::new(HashMap::new()));
         let spec = room_spec();
         let caller_fid =
             StatefulFid::try_from("pub-sub-video-call/caller".to_string())
                 .unwrap();
 
-        actix::spawn(test_for_create!(
+        test_for_create!(
             room_service,
             CreateRoom { spec },
             caller_fid,
             |member_el| {
                 assert_eq!(member_el.get_member().get_pipeline().len(), 1);
             }
-        ));
-
-        sys.run().unwrap();
+        )
+        .await;
     }
 
-    #[test]
-    fn create_member() {
-        let sys = actix::System::new("room-service-tests");
-
+    #[actix_rt::test]
+    async fn create_member() {
         let spec = room_spec();
         let member_spec = spec
             .members()
@@ -725,31 +725,28 @@ mod room_service_specs {
 
         let member_parent_fid = Fid::<ToRoom>::new(room_id);
         let member_id: MemberId = "test-member".to_string().into();
-        let member_fid: StatefulFid = member_parent_fid
+        let member_full_id: StatefulFid = member_parent_fid
             .clone()
             .push_member_id(member_id.clone())
             .into();
 
-        actix::spawn(test_for_create!(
+        test_for_create!(
             room_service,
             CreateMemberInRoom {
                 id: member_id,
                 spec: member_spec,
                 parent_fid: member_parent_fid,
             },
-            member_fid,
+            member_full_id,
             |member_el| {
                 assert_eq!(member_el.get_member().get_pipeline().len(), 1);
             }
-        ));
-
-        sys.run().unwrap();
+        )
+        .await;
     }
 
-    #[test]
-    fn create_endpoint() {
-        let sys = actix::System::new("room-service-tests");
-
+    #[actix_rt::test]
+    async fn create_endpoint() {
         let spec = room_spec();
 
         let mut endpoint_spec = spec
@@ -771,28 +768,27 @@ mod room_service_specs {
         let endpoint_parent_fid =
             Fid::<ToMember>::new(room_id, "caller".to_string().into());
         let endpoint_id: EndpointId = "test-publish".to_string().into();
-        let endpoint_fid: StatefulFid = endpoint_parent_fid
+        let endpoint_full_id: StatefulFid = endpoint_parent_fid
             .clone()
             .push_endpoint_id(endpoint_id.clone())
             .into();
 
-        actix::spawn(test_for_create!(
+        test_for_create!(
             room_service,
             CreateEndpointInRoom {
                 id: endpoint_id,
                 spec: endpoint_spec,
                 parent_fid: endpoint_parent_fid,
             },
-            endpoint_fid,
+            endpoint_full_id,
             |endpoint_el| {
                 assert_eq!(
                     endpoint_el.get_webrtc_pub().get_p2p(),
                     P2pMode::Never.into()
                 );
             }
-        ));
-
-        sys.run().unwrap();
+        )
+        .await;
     }
 
     /// Returns [`Future`] used for testing of all delete/get methods of
@@ -803,47 +799,38 @@ mod room_service_specs {
     /// deleted element is error then test considers successful.
     ///
     /// This function automatically stops [`actix::System`] when test completed.
-    fn test_for_delete_and_get(
+    async fn test_for_delete_and_get(
         room_service: Addr<RoomService>,
         element_fid: StatefulFid,
-    ) -> impl Future<Item = (), Error = ()> {
+    ) {
         let mut delete_msg = DeleteElements::new();
         delete_msg.add_fid(element_fid.clone());
         let delete_msg = delete_msg.validate().unwrap();
 
-        room_service
-            .send(delete_msg)
-            .and_then(move |res| {
-                res.unwrap();
-                room_service.send(Get(vec![element_fid]))
-            })
-            .map(move |res| {
-                assert!(res.is_err());
-                actix::System::current().stop();
-            })
-            .map_err(|e| panic!("{:?}", e))
+        room_service.send(delete_msg).await.unwrap().unwrap();
+        let get_result =
+            room_service.send(Get(vec![element_fid])).await.unwrap();
+
+        assert!(get_result.is_err());
+
+        actix::System::current().stop();
     }
 
-    #[test]
-    fn delete_and_get_room() {
-        let sys = actix::System::new("room-service-tests");
-
+    #[actix_rt::test]
+    async fn delete_and_get_room() {
         let room_id: RoomId = "pub-sub-video-call".to_string().into();
-        let room_fid = StatefulFid::from(Fid::<ToRoom>::new(room_id.clone()));
+        let room_full_id =
+            StatefulFid::from(Fid::<ToRoom>::new(room_id.clone()));
 
         let room_service = room_service(RoomRepository::new(hashmap!(
             room_id => Room::new(&room_spec(), &app_ctx()).unwrap().start(),
         )));
 
-        actix::spawn(test_for_delete_and_get(room_service, room_fid));
-
-        sys.run().unwrap();
+        test_for_delete_and_get(room_service, room_full_id).await;
     }
 
-    #[test]
-    fn delete_and_get_member() {
-        let sys = actix::System::new("room-service-tests");
-
+    #[actix_rt::test]
+    async fn delete_and_get_member() {
         let room_id: RoomId = "pub-sub-video-call".to_string().into();
         let member_fid = StatefulFid::from(Fid::<ToMember>::new(
             room_id.clone(),
@@ -854,15 +841,11 @@ mod room_service_specs {
             room_id => Room::new(&room_spec(), &app_ctx()).unwrap().start(),
         )));
 
-        actix::spawn(test_for_delete_and_get(room_service, member_fid));
-
-        sys.run().unwrap();
+        test_for_delete_and_get(room_service, member_fid).await;
     }
 
-    #[test]
-    fn delete_and_get_endpoint() {
-        let sys = actix::System::new("room-service-tests");
-
+    #[actix_rt::test]
+    async fn delete_and_get_endpoint() {
         let room_id: RoomId = "pub-sub-video-call".to_string().into();
         let endpoint_fid = StatefulFid::from(Fid::<ToEndpoint>::new(
             room_id.clone(),
@@ -874,8 +857,6 @@ mod room_service_specs {
             room_id => Room::new(&room_spec(), &app_ctx()).unwrap().start(),
         )));
 
-        actix::spawn(test_for_delete_and_get(room_service, endpoint_fid));
-
-        sys.run().unwrap();
+        test_for_delete_and_get(room_service, endpoint_fid).await;
     }
 }
